requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pymysql>=1.1.0
cryptography>=41.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0

# Desktop Wrappers (choose one or both)
Eel>=0.16.0              # Lightweight desktop wrapper using WebSocket bridge
pywebview>=4.0           # Native desktop UI using WebKit/CEF

# Additional utilities
requests>=2.31.0         # For API calls if needed
Pillow>=9.0.0            # Image processing for profile images

